Extract task run finalization logic into dedicated module
